Moves the InvokeContext dyn Trait and the ProcessInstructionWithContext type to invoke_context.rs
