+++ programs/compute-budget/src/lib.rs
@@ -1,4 +1,5 @@
-use solana_sdk::{instruction::InstructionError, process_instruction::InvokeContext};
+use solana_program_runtime::invoke_context::InvokeContext;
+use solana_sdk::instruction::InstructionError;
 
 pub fn process_instruction(
     _first_instruction_account: usize,
