+++ runtime/src/bank.rs
@@ -72,7 +72,7 @@ use solana_metrics::{inc_new_counter_debug, inc_new_counter_info};
 use solana_program_runtime::{
     instruction_processor::{ExecuteDetailsTimings, Executor, Executors, InstructionProcessor},
     instruction_recorder::InstructionRecorder,
-    invoke_context::ComputeMeter,
+    invoke_context::{ComputeMeter, ProcessInstructionWithContext},
     log_collector::LogCollector,
 };
 #[allow(deprecated)]
@@ -110,7 +110,6 @@ use solana_sdk::{
     nonce, nonce_account,
     packet::PACKET_DATA_SIZE,
     precompiles::get_precompiles,
-    process_instruction::ProcessInstructionWithContext,
     program_utils::limited_deserialize,
     pubkey::Pubkey,
     secp256k1_program,
@@ -6470,6 +6469,7 @@ pub(crate) mod tests {
         status_cache::MAX_CACHE_ENTRIES,
     };
     use crossbeam_channel::{bounded, unbounded};
+    use solana_program_runtime::invoke_context::InvokeContext;
     #[allow(deprecated)]
     use solana_sdk::sysvar::fees::Fees;
     use solana_sdk::{
@@ -6485,7 +6485,6 @@ pub(crate) mod tests {
         message::{Message, MessageHeader},
         nonce,
         poh_config::PohConfig,
-        process_instruction::InvokeContext,
         rent::Rent,
         signature::{keypair_from_seed, Keypair, Signer},
         stake::{
