+++ runtime/src/system_instruction_processor.rs
@@ -1,6 +1,6 @@
 use crate::nonce_keyed_account::NonceKeyedAccount;
 use log::*;
-use solana_program_runtime::ic_msg;
+use solana_program_runtime::{ic_msg, invoke_context::InvokeContext};
 use solana_sdk::{
     account::{AccountSharedData, ReadableAccount, WritableAccount},
     account_utils::StateMut,
@@ -8,7 +8,6 @@ use solana_sdk::{
     instruction::InstructionError,
     keyed_account::{from_keyed_account, get_signers, keyed_account_at_index, KeyedAccount},
     nonce,
-    process_instruction::InvokeContext,
     program_utils::limited_deserialize,
     pubkey::Pubkey,
     system_instruction::{NonceError, SystemError, SystemInstruction, MAX_PERMITTED_DATA_LENGTH},
